+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to `lua-vips` will be documented in this file.
 # 1.1-4 - 2017-08-30
 
 - small doc fixes
+- fix get() on gobject enum properties with older libvips
+- test for gobject enum properties as strings
 
 # 1.1-3 - 2017-08-08